+++ packages/collector/src/actions/profiling/cpu.js
@@ -23,8 +23,7 @@ exports.init = function() {
         'Instana for this application. This typically occurs when native addons could not be ' +
         'installed during module installation (npm install). See the instructions to learn more ' +
         'about the requirements of the collector: ' +
-        // eslint-disable-next-line max-len
-        'https://github.com/instana/nodejs-sensor/blob/master/packages/collector/README.md#cpu-profiling-garbage-collection-and-event-loop-information'
+        'https://www.instana.com/docs/ecosystem/node-js/installation/#native-addons'
     );
   }
 };