+++ packages/legacy-sensor/src/index.js
@@ -6,8 +6,7 @@ var instanaNodeJsCollector = require('@instana/collector');
 console.warn(
   'The package instana-nodejs-sensor has been renamed to @instana/collector. Please replace the dependency ' +
     '"instana-nodejs-sensor" in your package.json file with "@instana/collector". ' +
-    // eslint-disable-next-line max-len
-    'See https://github.com/instana/nodejs-sensor/blob/master/packages/collector/README.md#migrating-from-instana-nodejs-sensor-to-instanacollector for details.'
+    'See https://www.instana.com/docs/ecosystem/node-js/installation/#change-of-package-name for details.'
 );
 
 module.exports = exports = instanaNodeJsCollector;